chore: sync upstream rolldown v1.2.6
Upstream commit: c1767fcb88fae8e4d91e0ff29318d69bbb1afdea

OXC react compiler pin: crates_v0.147.0#4e258430cdb290598d9f2aeb2d13be598ec9e8e9

+++ crates/rolldown_binding/src/types/js_regex.rs
@@ -14,7 +14,7 @@ use rolldown_utils::js_regex::HybridRegex;
 pub struct JsRegExp {
   pub source: String,
   pub flags: String,
-  pub last_index: usize,
+  pub last_index: u32,
 }
 
 impl ValidateNapiValue for JsRegExp {}
@@ -40,7 +40,7 @@ impl FromNapiValue for JsRegExp {
     if js_object.instanceof(regexp_constructor)? {
       let source = js_object.get_named_property::<String>("source")?;
       let flags = js_object.get_named_property::<String>("flags")?;
-      let last_index = js_object.get_named_property::<u32>("lastIndex").unwrap_or(0) as usize;
+      let last_index = js_object.get_named_property::<u32>("lastIndex").unwrap_or(0);
 
       Ok(JsRegExp { source, flags, last_index })
     } else {
